first phase of 64 bit compatibility effort: the special1 and special2 fields of the mobj_t structure must be intptr_t, not int, they may be used to store a state OR to store the address of an mobj_t or player_t structure. the save system is still broken with 64 bit arches, it needs more work.
git-svn-id: svn://svn.code.sf.net/p/hhexen/code/trunk/hhexen@200 989aa981-fe4f-0410-aa6a-c19d413bcdcd

+++ base/a_action.c
@@ -4,8 +4,8 @@
 //** a_action.c : Heretic 2 : Raven Software, Corp.
 //**
 //** $RCSfile: a_action.c,v $
-//** $Revision: 1.5 $
-//** $Date: 2008-06-26 09:52:27 $
+//** $Revision: 1.6 $
+//** $Date: 2008-07-07 09:55:18 $
 //** $Author: sezero $
 //**
 //**************************************************************************
@@ -1077,7 +1077,7 @@ void A_ThrustInitDn(mobj_t *actor)
 	actor->flags = 0;
 	actor->flags2 = MF2_NOTELEPORT|MF2_FLOORCLIP|MF2_DONTDRAW;
 	mo = P_SpawnMobj(actor->x, actor->y, actor->z, MT_DIRTCLUMP);
-	actor->special1 = (int)mo;
+	actor->special1 = (intptr_t)mo;
 }
 
 void A_ThrustRaise(mobj_t *actor)
